Remove most dependencies on the worktree.
This is a major refactoring. Moving from context
to worktree as the link back to origins was not
enough as the worktree can be modified, so we
still needed to copy, or conjure up a worktree
when all we had was a repository.

Both worktrees and repositories now support git
commands executed in their context, with the
repository preferred when both are available.

Various outputs are improved, and the
committer/author/tagger + date stuff is now
handled by Person and CommittedBy classes rather
than multiple properties for each. This simplifies
the model and it is now easy to collect all the
users in a repository, for example.

We no longer use the xonsh subprocess code, as it
does not expose setting the working directory for
commands. Instead, we use the subprocess module
directly, which is more flexible but does not
integrate with the tracing etc. provided by xonsh.