Remove  module-unloading from tests.
It helped with catching loading issues, but when run under poetry, the tests can encounter different versions of themselves. The primary symptom is try/catch fails to catch what ought be caught.
It's all for naught.

We have very limited global state (event handlers, etc.), and we've expunged tests which depend on those, so it's load once and go hence and try not to break anything in the test repo.

But this reveals that many tests are sensitive to the detached state of the CI repo, rather than operating entirely in their own test repo.

This is, unfortunately, out of our control, as it happens at code loading/initialization time.

We'll have to add detached state tests for the code  that  runs at load time (it's not specific to load time), to be sure to catch these before sending off to CI.